consumer/requirements.txt
pika==1.2.0
Faker==8.10.1
python-dotenv==0.18.0